requirements.txt
bcrypt==3.2.0
fastapi==0.78.0
passlib==1.7.4
pydantic==1.9.1
pytest==6.2.1
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
SQLAlchemy==1.3.22
starlette==0.19.1
uvicorn==0.13.3
fastapi-jwt-auth==0.5.0
aiofiles==0.8.0